% Generated by roxygen2 (4.0.2): do not edit by hand
\name{l1tf}
\alias{l1tf}
\title{L1 trend filteirng}
\usage{
l1tf(x, lambda = 0.1)
}
\arguments{
\item{x}{Numeric vector of inputs. Must be equally spaced in time.}

\item{lambda}{A number that controls the balance between accuracy and
simplicity. Small values favour accuracy; large values favour simplicity.}
}
\description{
L1 trend filteirng
}
\examples{
x <- runif(1000)

plot(x)
lines(l1tf(x, 0.5))
}

