% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{snp500}
\alias{snp500}
\title{SNP 500 data}
\format{\preformatted{ num [1:2000] 7.16 7.18 7.17 7.16 7.17 ...
}}
\usage{
snp500
}
\description{
The Daily Standard & Poor index from March 25, 1999 to March 9, 2007.
}
\keyword{datasets}

