% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp500}
\alias{sp500}
\title{SP 500 data}
\format{
A data frame with 2001 rows and 3 columns:
\describe{
\item{date}{date of measurement}
\item{raw}{Raw value of SP500}
\item{log}{Logged value, as used in paper}
}
}
\usage{
sp500
}
\description{
The Daily Standard & Poor index from March 25, 1999 to March 9, 2007.
}
\examples{
plot(sp500$date, sp500$log, type = "l")
lines(sp500$date, l1tf(sp500$log, lambda = 50), col = "red")
lines(sp500$date, l1tf(sp500$log, prop = 0.01), col = "blue")
}
\keyword{datasets}
